==> code/4-prelim_markdown.Rmd <==
---
title: "preliminary results"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=FALSE,message=FALSE,warning=FALSE,
                      collapse = TRUE,
                      comment = "#>", 
                      fig.path = "markdown/prelim/"
                      )
```

```{r script setup}
library(here)
source(here("code/0-packages.R"))
theme_set(theme_bw())
```

final cores F31, F32 removed

----

### Hypotheses:
(1) increased **temperature** would destabilize MAOM, increasing the bioavailable DOM and C mineralization (also increase microbial activity); 
(2) increased **moisture content** would improve hydrologic connectivity, increasing CO2 flux; and 
(3) higher temperature would have a greater destabilization effect on SOM than higher moisture, because ---. 
(4) Since sorption to mineral surfaces is an important/dominant mechanism of SOM stabilization, we also hypothesized that  destabilization of SOM due to increased temperature and/or moisture would be less in **clay-rich soils**. 

----

----

# RESPIRATION

- cumulative CO2 evolution was influenced by temperature and clay amendments, but not by moisture content
- CO2 evolution was greater at 20 C than at 4 C
- CO2 evolution was reduced in clay-amended soils. this effect was seen only at 20 C

----

```{r flux_files}
flux = read.csv(here("data/processed/flux.csv")) %>% 
  mutate(date = ymd(date),
         Moisture = factor(Moisture, levels = c("50% WFPS", "100% WFPS")),
         CO2C_ug_g_cum = CO2C_ug_cum/42)
```

## core-wise graphs

**CO2-C concentrations**

```{r resp_conc_graph, fig.width=10}
flux %>% 
  ggplot(aes(x = date, y = CO2_1_ppm, color = as.character(core)))+
  geom_point()+ geom_path()+
  #geom_point(data = flux, aes(y = CO2_2_ppm), color = "black")+
  facet_grid(Moisture ~ Clay+Temperature)+
  theme_kp()+
  theme(legend.position = "none",
        axis.text.x = element_text(size=10))+
  NULL
```

**CO2-C fluxes**

```{r flux_cores}
flux %>% 
  ggplot(aes(x = incub_days, y = CO2C_ug_g_hr, color = as.character(core)))+
  geom_point()+geom_path()+
  facet_grid(Moisture ~ Clay+Temperature)+
  theme(legend.position = "none")+
  NULL
```

**cumulative CO2-C evolved**

```{r flux_cum}
flux %>% 
  ggplot(aes(x = incub_days, y = CO2C_ug_cum, color = as.character(core)))+
  geom_point()+geom_path()+
  facet_grid(Moisture ~ Clay+Temperature)+
  theme(legend.position = "none")+
  NULL
```

note: some cores are missing time-points. missing values must be filled in with interpolation


## mean/summary graphs

```{r flux_means, fig.width=6, fig.height=6}

labels = tribble(
  ~Temperature, ~Moisture, ~Clay,
  4, "50% WFPS", "non-amended"
)

flux %>% 
  group_by(Temperature, Clay, Moisture, incub_days) %>% 
  summarize(CO2C_ug_g_hr = mean(CO2C_ug_g_hr, na.rm = T)) %>% 
  arrange(incub_days) %>% 
  ggplot(aes(x = incub_days, y = CO2C_ug_g_hr, color = Clay))+
  geom_path()+
#  geom_smooth()+
  geom_point(data = flux, size=1.5, alpha = 0.3)+
  scale_color_manual(values = soilpalettes::soil_palette("eutrostox",2))+
  geom_text(data = labels, aes(x = 3, y = 0.12), label = "time zero\n (initial)", color = "black", size=3)+
  geom_curve(data = labels, aes(x = 0.3, xend = 1.8, y = 0.11, yend = 0.12), curvature = -0.3, color = "black")+
  labs(title = "mean CO2-C flux",
       y = "CO2-C, ug/g/hr",
       x = "days of incubation")+
  facet_grid(Moisture ~ Temperature)+
  theme_kp()+
  NULL
```

```{r flux_cum_means, fig.width=6, fig.height=6}
flux %>% 
  group_by(Temperature, Clay, Moisture, incub_days) %>% 
  summarize(CO2C_ug_g_cum = mean(CO2C_ug_g_cum)) %>% 
  arrange(incub_days) %>% 
  ggplot(aes(x = incub_days, y = CO2C_ug_g_cum, color = Clay))+
  geom_path()+
#  geom_smooth()+
  geom_point(data = flux, size=1.5, alpha = 0.3)+
  scale_color_manual(values = soilpalettes::soil_palette("eutrostox",2))+
  labs(title = "cumulative CO2-C evolved",
       y = "CO2-C, ug/g",
       x = "days of incubation")+
  facet_grid(Moisture ~ Temperature)+
  theme_kp()+
  NULL
```




## stats

```{r}
cum_flux = 
  flux %>% 
  group_by(core) %>% 
  filter(CO2C_ug_cum == max(CO2C_ug_cum))
```


```{r flux_aov, echo=TRUE, fig.width=4}
(cum_flux_aov = car::Anova(lm(
  log10(CO2C_ug_cum) ~ (Temperature + Clay + Moisture)^2,
  data = cum_flux), 
  type = "III")
)
```

**cumulative CO2-C evolved (ug/g)**
```{r flux_cum_table}
flux %>% 
  group_by(core, Temperature, Clay, Moisture) %>% 
  summarize(CO2C_ug_g_cum = max(CO2C_ug_g_cum)) %>% 
  group_by(Temperature, Clay, Moisture) %>% 
  summarise(CO2C_ug_g_cum_mean = mean(CO2C_ug_g_cum),
            se_CO2C = sd(CO2C_ug_g_cum, na.rm = TRUE)/sqrt(n()),
            CO2C_ug_g_cum = paste(round(CO2C_ug_g_cum_mean,2), "\u00b1", round(se_CO2C,2))) %>% 
  select(Temperature, Clay, Moisture, CO2C_ug_g_cum) %>% 
  spread(Clay, CO2C_ug_g_cum) %>% 
  knitr::kable()
```

---

# DOC

- DOC for initial soils was influenced by moisture content and clay amendment, but not by temperature (the soils were air-dried, which may have caused destabilization of protected C -- and that may have masked any temperature effects)
- after the 7-day incubation, DOC was influenced by moisture, but not by clay amendment or temperature


```{r doc_files}
doc = read.csv(here("data/processed/doc.csv"))
```


```{r doc_scatter, fig.width=10, fig.height=4}
doc %>%
  mutate(Temp_C = factor(Temp_C, levels = c(4, 20))) %>% 
  mutate(Time = factor(Time, levels = c("Initial", "Final"))) %>% 
  ggplot(aes(x = Temp_C, y = doc_ppm, shape = interaction(Clay, Moisture), color = Moisture))+
  geom_point(size=3, stroke=1, position = position_dodge(width = 0.6))+
  scale_shape_manual(values = c(1,16,0,15))+
  scale_color_manual(values = rev(soilpalettes::soil_palette("redox2",2)))+
  
  labs(x = "Temperature, C",
       y = "WEOC, ppm",
       shape = "",
       color = "")+
  facet_grid(.~Time)+
  theme_kp()+
  theme(legend.position = "right")+
  NULL
```

## summary table

```{r doc_table}
doc %>% 
  mutate(Time = factor(Time, levels = c("Initial", "Final"))) %>% 
  select(doc_ppm, Time, Temp_C, Clay, Moisture) %>% 
  group_by(Time, Temp_C, Clay, Moisture) %>% 
  summarise(doc = mean(doc_ppm),
            se_doc = sd(doc_ppm)/sqrt(n()),
            mean = paste(round(doc,2), "\u00b1", round(se_doc,2))) %>% 
  select(-doc, -se_doc) %>% 
  spread(Clay, mean) %>% 
  knitr::kable()
```

<details>
  <summary>stats</summary>

```{r doc_stats, echo=TRUE}
aov_doc_initial = aov(doc_ppm ~ (Temp_C + Moisture + Clay)^2, 
                    data = doc %>% filter(Time=="Initial"))
summary(aov_doc_initial)

aov_doc_final = aov(doc_ppm ~ (Temp_C + Moisture + Clay)^2, 
                    data = doc %>% filter(Time=="Final"))
summary(aov_doc_final)
```

</details>


---

<details>
  <summary>Session Info</summary>

date run: `r Sys.Date()`

```{r}
sessionInfo()
```

</details>
